+++ app/src/main/java/edu/agh/susgame/front/model/graph/Edge.kt
@@ -19,4 +19,5 @@ class Edge(
     fun removePlayer(playerId: PlayerId){
         playersIdsUsingEdge.remove(playerId)
     }
+
 }
