+++ app/src/main/java/edu/agh/susgame/front/service/mock/MockServerMapProvider.kt
@@ -1,6 +1,5 @@
 package edu.agh.susgame.front.providers.mock
 
-import androidx.compose.ui.graphics.Color
 import edu.agh.susgame.dto.rest.model.LobbyId
 import edu.agh.susgame.dto.rest.model.Player
 import edu.agh.susgame.dto.rest.model.PlayerId
@@ -66,17 +65,17 @@ class MockServerMapProvider(mockDelayMs: Long? = null) : ServerMapProvider {
         val players = listOf(
             Player(
                 nickname = PlayerNickname("Player_0"),
-                colorHex = Color.Red.value.toLong(),
+                colorHex = 0xAF00FF00,
                 id = PlayerId(0)
             ),
             Player(
                 nickname = PlayerNickname("Player_1"),
-                colorHex = Color.Blue.value.toLong(),
+                colorHex = 0xAFF000FF,
                 id = PlayerId(1)
             ),
             Player(
                 nickname = PlayerNickname("Player_2"),
-                colorHex = Color.Magenta.value.toLong(),
+                colorHex = 0xAFFF5733,
                 id = PlayerId(2)
             ),
         )
