+++ app/src/main/java/edu/agh/susgame/front/ui/components/game/navigation/GameNavBarComp.kt
@@ -1,0 +1,41 @@
+package edu.agh.susgame.front.ui.components.game.navigation
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.rememberNavController
+import edu.agh.susgame.dto.rest.model.LobbyId
+import edu.agh.susgame.front.service.interfaces.GameService
+import edu.agh.susgame.front.service.interfaces.ServerMapProvider
+
+
+@Composable
+fun GameNavBarComp(
+    lobbyId: LobbyId?,
+    menuNavController: NavHostController,
+    serverMapProvider: ServerMapProvider,
+    gameService: GameService,
+) {
+    val gameNavController = rememberNavController()
+    Scaffold(
+        modifier = Modifier
+            .fillMaxSize(),
+//            .padding(top = PaddingL),
+        bottomBar = {
+            BottomBarComponent(gameNavController)
+        }
+    ) { padding ->
+        GameNavigationHost(
+            lobbyId,
+            padding,
+            menuNavController,
+            gameNavController,
+            serverMapProvider,
+            gameService,
+        )
+    }
+}
+
+
